+++ frontend/index.html
@@ -1,0 +1,16 @@
+<!doctype html>
+<html lang="en">
+
+<head>
+	<meta charset="UTF-8" />
+	<link rel="icon" type="image/svg+xml" href="/vite.svg" />
+	<meta name="viewport" content="width=device-width, initial-scale=1.0" />
+	<title>Pong</title>
+</head>
+
+<body class="font-monospace">
+	<div id="root"></div>
+	<script type="module" src="/src/main.jsx"></script>
+</body>
+
+</html>
